Update the IMS UCE and CM HAL versions to 2.0
On older modems uceservice and cmservice HAL 2.0 is
supported. Also new IMS HALs like callinfo and
imsfactory are not supported older targets. This
causes VTS to fail since HAL registration is not
successful at boot up when hybrid NV is disabled.

Update entry of uceservice and cmservice HAL to 2.0
and remove entries for ImsFactory and callinfo HAL.

Change-Id: Iefb1630690f516c7ae08580b57fd425981f2e934